tests: Split unit and integration RLP tests
